proceed even if input journal(s) have errors
Log by default (at warning level) the number of errors encountered in input
ledgers; log with increased verbosity (info level) the details of all
encountered errors.

Fail with an error if no valid entry is found after loading.

This strikes a good compromise between being lenient in the face of errors that
can be present and still failing with exit status 2 if nothing has been parsed.

Closes #5