Proof of concept for discussion:
This allows the beancount file to be specified
via an environment variable BEANCOUNT_FILE, so
you don't need to repeat it every time you run
bean-grep.

As part of this change, the PATTERN is now
mandatory. This is because I don't
really understand how to make click keep working
the old way while also allowing the filename
to come via an environment variable. We could
pull the environment variable in ourselves manually
but it feels cleaner to let click do that.

I'm not sure if there's really a usecase for not
having a pattern, anyway? That's just cat. I don't
think regular grep allows no pattern?